requirements.txt
asgiref==3.5.2
Django==4.0.4
gunicorn==20.1.0
psycopg2==2.8.6
sqlparse==0.4.2
backports.zoneinfo==0.2.1
Pillow==9.1.0
python-slugify==6.1.2
text-unidecode==1.3
whitenoise==6.0.0
pytz==2021.1
mysqlclient>=2.0
django-mptt==0.14.0
gunicorn==20.1.0